requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
numpy==2.3.5
PyYAML==6.0.3
httpx==0.28.1
torch==2.6.0
torchaudio==2.6.0
speechbrain==1.0.3
huggingface_hub==1.7.1
